refactor(db): move search proptests + delete in-source mod tests
WHY: Batch G cluster 3 of 3 (final). Both proptests
(fts_consistent_under_tag_churn + fts_matches_ground_truth_under_soft_delete_churn)
move to search_proptests.rs with case counts unchanged (#124 cap preserved).
Now-empty #[cfg(test)] mod tests block deleted from src/search_repo.rs.
Stale comments at fts_codegen_round_trip.rs:~57,~205 updated to point at
common::seed_conn / common::test_db. Long dead_code WHY in common/mod.rs
wrapped to ≤100-char lines (Task 4 review NIT).

Production search_repo.rs settles at ~120 LOC; audit §A9's literal 4-way
split (mod.rs/filters.rs/query.rs/rows.rs) declared SUPERSEDED by Batch C —
no meaningful production split surface remains. Test count unchanged (118).
